requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.7
django-mptt==0.14.0
django-oauth-toolkit==1.7.1
python-decouple==3.8
celery==5.3.4
redis==5.0.1
django-cors-headers==4.3.1
requests==2.31.0
coverage==7.3.2
pytest==7.4.3
pytest-django==4.5.2
pytest-cov==4.1.0
gunicorn==21.2.0
whitenoise==6.6.0
django-extensions==3.2.3
drf-yasg==1.21.7
Pillow==10.1.0